1st
A character device is one that can be accessed as a stream of bytes; a
character device driver is a module in charge of implementation this
behaviour. Such as driver usually implements at least the open, close,
read, and write system calls. Character devices are accessed by means
of file system nodes such as /dev/tty1 and /dev/lp0

character device driver (a module) with functions to open, close, read,
and write a device.

Develop a user application to write and read message to the device.